LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-hover-expected.txt



FAIL CSS Test: ::scroll-marker changes on hover via :hover assert_equals: ::scroll-marker changes on hover via :hover expected "rgb(0, 0, 255)" but got ""
